added sortBar
